5. Подсчет функции в диапазоне. [#24251 #226637]

+++ src/main/java/ru/job4j/lambda/FuncDiapason.java
@@ -7,8 +7,8 @@ import java.util.function.*;
 public class FuncDiapason {
     public static List<Double> diapason(int start, int end, Function<Double, Double> func) {
         List<Double> result = new ArrayList<>();
-        for (double i = start; i < end; i++) {
-            result.add(func.apply(i));
+        for (int i = start; i < end; i++) {
+            result.add(func.apply((double) i));
         }
         return result;
     }
